Rename Dashboard to PublicController and remove previous demo route
