+++ src/main/java/com/bytebreeze/quickdrop/controller/PublicController.java
@@ -1,0 +1,20 @@
+package com.bytebreeze.quickdrop.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+
+@Controller
+@RequestMapping
+public class PublicController {
+
+    @GetMapping("/")
+    public String home(Model model) {
+        model.addAttribute("title", "Dashboard - Home");
+        return "index";  // Removed leading slash
+    }
+
+
+}
